dataLengthToDlc and dlcToDataLength use uint16_t instead of uint8_t

+++ canard_internals.h
@@ -85,8 +85,8 @@ CANARD_INTERNAL void incrementTransferID(uint8_t* transfer_id);
 CANARD_INTERNAL uint64_t releaseStatePayload(CanardInstance* ins,
                                              CanardRxState* rxstate);
 
-CANARD_INTERNAL uint8_t dlcToDataLength(uint8_t dlc);
-CANARD_INTERNAL uint8_t dataLengthToDlc(uint8_t data_length);
+CANARD_INTERNAL uint16_t dlcToDataLength(uint16_t dlc);
+CANARD_INTERNAL uint16_t dataLengthToDlc(uint16_t data_length);
 
 /// Returns the number of frames enqueued
 CANARD_INTERNAL int16_t enqueueTxFrames(CanardInstance* ins,
